+++ src/modules/stock/stock.controller.ts
@@ -9,6 +9,11 @@ export class StockController {
     try {
       const businessId = req.user?.business_id;
 
+      if (!businessId)
+        return res
+          .status(400)
+          .json({ success: false, message: "Business ID missing", data: {} });
+
       const data = await this.service.saveStock(businessId!, req.body);
 
       res.json({
@@ -24,19 +29,57 @@ export class StockController {
     }
   };
 
+  assignStockLocations = async (req: AuthRequest, res: Response) => {
+    try {
+      const businessId = req.user?.business_id;
+      const { stockId } = req.params;
+
+      await this.service.assignStockLocations(
+        Number(stockId),
+        businessId!,
+        req.body,
+      );
+
+      res.json({
+        success: true,
+        message: "Storage assigned",
+      });
+    } catch (error: any) {
+      res.status(400).json({
+        success: false,
+        message: error.message,
+      });
+    }
+  };
+
   getStocks = async (req: AuthRequest, res: Response) => {
+    const userId = req.user!.id;
+
     const businessId = req.user?.business_id;
 
-    const data = await this.service.getStocks(businessId!);
+    if (!businessId)
+      return res
+        .status(400)
+        .json({ success: false, message: "Business ID missing", data: {} });
+
+    const data = await this.service.getStocks(userId, businessId!);
 
     res.json({ success: true, data });
   };
 
   getStockById = async (req: AuthRequest, res: Response) => {
+    const userId = req.user!.id;
+
     const businessId = req.user?.business_id;
+
+    if (!businessId)
+      return res
+        .status(400)
+        .json({ success: false, message: "Business ID missing", data: {} });
 
     const data = await this.service.getStockById(
       Number(req.params.id),
+      userId,
       businessId!,
     );
 
@@ -46,6 +89,11 @@ export class StockController {
   deleteStock = async (req: AuthRequest, res: Response) => {
     const businessId = req.user?.business_id;
 
+    if (!businessId)
+      return res
+        .status(400)
+        .json({ success: false, message: "Business ID missing", data: {} });
+
     await this.service.deleteStock(Number(req.params.id), businessId!);
 
     res.json({ success: true, message: "Stock deleted" });
@@ -54,6 +102,11 @@ export class StockController {
   updateStock = async (req: AuthRequest, res: Response) => {
     try {
       const businessId = req.user?.business_id;
+      if (!businessId)
+        return res
+          .status(400)
+          .json({ success: false, message: "Business ID missing", data: {} });
+
       const stockId = Number(req.params.id);
 
       const data = await this.service.updateStock(
@@ -72,6 +125,70 @@ export class StockController {
         success: false,
         message: error.message,
       });
+    }
+  };
+
+  getAssignedLocations = async (req: AuthRequest, res: Response) => {
+    try {
+      const businessId = req.user?.business_id;
+      const { stockId } = req.params;
+
+      const data = await this.service.getAssignedLocations(
+        Number(stockId),
+        businessId!,
+      );
+
+      res.json({
+        success: true,
+        data,
+      });
+    } catch (err: any) {
+      res.status(500).json({
+        success: false,
+        message: err.message,
+      });
+    }
+  };
+
+  deleteAssignedLocations = async (req: AuthRequest, res: Response) => {
+    try {
+      const businessId = req.user?.business_id;
+      const { stockId } = req.params;
+
+      await this.service.deleteAssignedLocations(Number(stockId), businessId!);
+
+      res.json({
+        success: true,
+        message: "All storage assignments removed",
+      });
+    } catch (err: any) {
+      res.status(400).json({
+        success: false,
+        message: err.message,
+      });
+    }
+  };
+
+  updateSingleLocation = async (req: AuthRequest, res: Response) => {
+    try {
+      const businessId = req.user?.business_id;
+
+      await this.service.updateSingleLocation(req.body, businessId!);
+
+      res.json({ success: true });
+    } catch (err: any) {
+      res.status(400).json({ success: false, message: err.message });
+    }
+  };
+
+  deleteSingleLocation = async (req: AuthRequest, res: Response) => {
+    try {
+      const businessId = req.user?.business_id;
+      await this.service.deleteSingleLocation(req.body, businessId!);
+
+      res.json({ success: true });
+    } catch (err: any) {
+      res.status(400).json({ success: false, message: err.message });
     }
   };
 }
